Rely on well-known /usr/bin/env to find Ruby
This is pretty darn normalized on *nix systems these days, where this
will be installed via Homebrew. Let's rely on that rather than
hard-coding to a version of Ruby that can/will likely change over time.

Also, use the Regex#match? and avoid extra object; we don't need/use the
MatchData, so don't create it.